requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
fastapi==0.115.0
jinja2==3.1.4
qrcode==7.4.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart